+++ src/Patch/InstrRule.h
@@ -33,12 +33,66 @@ namespace llvm {
 
 namespace QBDI {
 
+class Assembly;
 class Patch;
 class PatchGenerator;
 
 /*! An instrumentation rule written in PatchDSL.
 */
-class InstrRule : public AutoAlloc<InstrRule, InstrRule> {
+class InstrRule {
+
+    protected:
+
+    // priority of the rule.
+    // The rule with the lesser priority will be applied first
+    int priority;
+
+    public:
+
+    InstrRule(int priority = 0) : priority(priority) {}
+
+    virtual ~InstrRule() = default;
+
+    virtual operator std::unique_ptr<InstrRule>() =0;
+
+    // virtual copy constructor used to duplicate the object
+    virtual std::unique_ptr<InstrRule> clone() const =0;
+
+    virtual RangeSet<rword> affectedRange() const =0;
+
+    inline int getPriority() const { return priority; };
+
+    inline void setPriority(int priority) { this->priority = priority; };
+
+    inline virtual void changeVMInstanceRef(VMInstanceRef vminstance) {};
+
+    /*! Determine wheter this rule have to be apply on this Path and instrument if needed.
+     *
+     * @param[in] patch     The current patch to instrument.
+     * @param[in] MCII      A LLVM::MCInstrInfo classes used for internal architecture specific
+     *                      queries.
+     * @param[in] MRI       A LLVM::MCRegisterInfo classes used for internal architecture specific
+     *                      queries.
+     * @param[in] assemby   Assembly object to generate InstAnalysis
+    */
+    virtual bool tryInstrument(Patch &patch, const llvm::MCInstrInfo* MCII, const llvm::MCRegisterInfo* MRI,
+                               const Assembly* assembly) const =0;
+
+    /*! Instrument a patch by evaluating its generators on the current context. Also handles the
+     *  temporary register management for this patch.
+     *
+     * @param[in] patch  The current patch to instrument.
+     * @param[in] MCII   A LLVM::MCInstrInfo classes used for internal architecture specific
+     *                   queries.
+     * @param[in] MRI    A LLVM::MCRegisterInfo classes used for internal architecture specific
+     *                   queries.
+    */
+    void instrument(Patch &patch, const llvm::MCInstrInfo* MCII, const llvm::MCRegisterInfo* MRI,
+                    const std::vector<std::shared_ptr<PatchGenerator>> patchGen, bool breakToHost, InstPosition position) const;
+};
+
+
+class InstrRuleBasic : public InstrRule {
 
     PatchCondition::SharedPtr                     condition;
     std::vector<std::shared_ptr<PatchGenerator>>  patchGen;
@@ -58,13 +112,23 @@ class InstrRule : public AutoAlloc<InstrRule, InstrRule> {
      * @param[in] breakToHost  A boolean determining whether this instrumentation should end with
      *                         a break to host (in the case of a callback for example).
     */
-    InstrRule(PatchCondition::SharedPtr condition, std::vector<std::shared_ptr<PatchGenerator>> patchGen,
-              InstPosition position, bool breakToHost) : condition(condition),
-              patchGen(patchGen), position(position), breakToHost(breakToHost) {}
+    InstrRuleBasic(PatchCondition::SharedPtr condition, std::vector<std::shared_ptr<PatchGenerator>> patchGen,
+                   InstPosition position, bool breakToHost, int priority = 0) : InstrRule(priority), condition(condition), patchGen(patchGen),
+        position(position), breakToHost(breakToHost) {}
 
-    InstPosition getPosition() const { return position; }
+    ~InstrRuleBasic() override = default;
 
-    RangeSet<rword> affectedRange() const {
+    inline operator std::unique_ptr<InstrRule>() override {
+        return std::make_unique<InstrRuleBasic>(*this);
+    }
+
+    inline std::unique_ptr<InstrRule> clone() const override {
+        return std::make_unique<InstrRuleBasic>(*this);
+    };
+
+    inline InstPosition getPosition() const { return position; }
+
+    inline RangeSet<rword> affectedRange() const override {
         return condition->affectedRange();
     }
 
@@ -78,16 +142,51 @@ class InstrRule : public AutoAlloc<InstrRule, InstrRule> {
     */
     bool canBeApplied(const Patch &patch, const llvm::MCInstrInfo* MCII) const;
 
-    /*! Instrument a patch by evaluating its generators on the current context. Also handles the
-     *  temporary register management for this patch.
-     *
-     * @param[in] patch  The current patch to instrument.
-     * @param[in] MCII   A LLVM::MCInstrInfo classes used for internal architecture specific
-     *                   queries.
-     * @param[in] MRI    A LLVM::MCRegisterInfo classes used for internal architecture specific
-     *                   queries.
-    */
-    void instrument(Patch &patch, const llvm::MCInstrInfo* MCII, const llvm::MCRegisterInfo* MRI) const;
+    inline bool tryInstrument(Patch &patch, const llvm::MCInstrInfo* MCII, const llvm::MCRegisterInfo* MRI,
+                              const Assembly* assembly) const override {
+        if (canBeApplied(patch, MCII)) {
+            instrument(patch, MCII, MRI, patchGen, breakToHost, position);
+            return true;
+        }
+        return false;
+    }
+};
+
+class InstrRuleUser : public InstrRule {
+
+    InstrumentCallback cbk;
+    AnalysisType analysisType;
+    void* cbk_data;
+    VMInstanceRef vm;
+    RangeSet<rword> range;
+
+    public:
+
+    InstrRuleUser(InstrumentCallback cbk, AnalysisType analysisType,
+                  void* cbk_data, VMInstanceRef vm, RangeSet<rword> range,
+                  int priority = 0) : InstrRule(priority),
+        cbk(cbk), analysisType(analysisType), cbk_data(cbk_data), vm(vm), range(range) {}
+
+    ~InstrRuleUser() override = default;
+
+    inline operator std::unique_ptr<InstrRule>() override {
+        return std::make_unique<InstrRuleUser>(*this);
+    }
+
+    inline std::unique_ptr<InstrRule> clone() const override {
+        return std::make_unique<InstrRuleUser>(*this);
+    };
+
+    inline void changeVMInstanceRef(VMInstanceRef vminstance) override {
+        vm = vminstance;
+    };
+
+    inline RangeSet<rword> affectedRange() const override {
+        return range;
+    }
+
+    bool tryInstrument(Patch &patch, const llvm::MCInstrInfo* MCII, const llvm::MCRegisterInfo* MRI,
+                       const Assembly* assembly) const override;
 };
 
 }