+++ test/API/VMTest.cpp
@@ -633,7 +633,8 @@ struct MoveCallbackStruct {
 
     bool reachEventCB;
     bool reachInstCB;
-    bool reachCB2 = false;
+    bool reachInstrumentCB;
+    bool reachCB2;
 };
 
 static QBDI::VMAction allowedNewBlock(QBDI::VMInstanceRef vm, const QBDI::VMState* state, QBDI::GPRState*, QBDI::FPRState*, void *data_) {
@@ -643,6 +644,15 @@ static QBDI::VMAction allowedNewBlock(QBDI::VMInstanceRef vm, const QBDI::VMStat
 
     data->reachEventCB = true;
     return QBDI::VMAction::CONTINUE;
+}
+
+static std::vector<QBDI::InstrumentDataCBK> instrumentCopyCB(QBDI::VMInstanceRef vm, const QBDI::InstAnalysis *inst, void* data_) {
+    MoveCallbackStruct* data = static_cast<MoveCallbackStruct*>(data_);
+    CHECK(data->expectedRef == vm);
+    CHECK(data->allowedNewBlock);
+
+    data->reachInstrumentCB = true;
+    return {};
 }
 
 static QBDI::VMAction verifyVMRef(QBDI::VMInstanceRef vm, QBDI::GPRState*, QBDI::FPRState*, void *data_) {
@@ -666,12 +676,13 @@ TEST_CASE("VMTest-MoveConstructor") {
     VMTest vm1;
     QBDI::VM* vm = vm1.vm.get();
 
-    MoveCallbackStruct data {vm, true, false, false};
+    MoveCallbackStruct data {vm, true, false, false, false, false};
 
     bool instrumented = vm->addInstrumentedModuleFromAddr((QBDI::rword)&dummyFunCall);
     REQUIRE(instrumented);
 
     vm->addCodeCB(QBDI::InstPosition::POSTINST, verifyVMRef, &data);
+    vm->addInstrRule(instrumentCopyCB, QBDI::ANALYSIS_INSTRUCTION, &data);
     vm->addVMEventCB(QBDI::SEQUENCE_ENTRY | QBDI::SEQUENCE_EXIT | QBDI::BASIC_BLOCK_NEW, allowedNewBlock, &data);
 
     QBDI::rword retvalue;
@@ -680,9 +691,11 @@ TEST_CASE("VMTest-MoveConstructor") {
     REQUIRE(retvalue == 350);
     REQUIRE(data.reachEventCB);
     REQUIRE(data.reachInstCB);
+    REQUIRE(data.reachInstrumentCB);
 
     data.reachEventCB = false;
     data.reachInstCB = false;
+    data.reachInstrumentCB = false;
     data.allowedNewBlock = false;
 
     vm = vm1.vm.release();
@@ -697,9 +710,16 @@ TEST_CASE("VMTest-MoveConstructor") {
     data.expectedRef = &movedVM;
 
     movedVM.call(&retvalue, (QBDI::rword) dummyFun1, {780});
+
     REQUIRE(retvalue == 780);
     REQUIRE(data.reachEventCB);
     REQUIRE(data.reachInstCB);
+    REQUIRE_FALSE(data.reachInstrumentCB);
+
+
+    data.allowedNewBlock = true;
+    movedVM.precacheBasicBlock((QBDI::rword) dummyFun0);
+    REQUIRE(data.reachInstrumentCB);
 }
 
 TEST_CASE("VMTest-CopyConstructor") {
@@ -707,12 +727,13 @@ TEST_CASE("VMTest-CopyConstructor") {
     VMTest vm1;
     QBDI::VM* vm = vm1.vm.get();
 
-    MoveCallbackStruct data {vm, true, false, false};
+    MoveCallbackStruct data {vm, true, false, false, false, false};
 
     bool instrumented = vm->addInstrumentedModuleFromAddr((QBDI::rword)&dummyFunCall);
     REQUIRE(instrumented);
 
     vm->addCodeCB(QBDI::InstPosition::POSTINST, verifyVMRef, &data);
+    vm->addInstrRule(instrumentCopyCB, QBDI::ANALYSIS_INSTRUCTION, &data);
     vm->addVMEventCB(QBDI::SEQUENCE_ENTRY | QBDI::SEQUENCE_EXIT | QBDI::BASIC_BLOCK_NEW, allowedNewBlock, &data);
 
     QBDI::rword retvalue;
@@ -721,9 +742,11 @@ TEST_CASE("VMTest-CopyConstructor") {
     REQUIRE(retvalue == 350);
     REQUIRE(data.reachEventCB);
     REQUIRE(data.reachInstCB);
+    REQUIRE(data.reachInstrumentCB);
 
     data.reachEventCB = false;
     data.reachInstCB = false;
+    data.reachInstrumentCB = false;
     data.allowedNewBlock = false;
 
     // copy vm
@@ -735,9 +758,11 @@ TEST_CASE("VMTest-CopyConstructor") {
     REQUIRE(retvalue == 620);
     REQUIRE(data.reachEventCB);
     REQUIRE(data.reachInstCB);
+    REQUIRE_FALSE(data.reachInstrumentCB);
 
     data.reachEventCB = false;
     data.reachInstCB = false;
+    data.reachInstrumentCB = false;
     data.allowedNewBlock = true;
     data.expectedRef = &movedVM;
 
@@ -745,6 +770,7 @@ TEST_CASE("VMTest-CopyConstructor") {
     REQUIRE(retvalue == 780);
     REQUIRE(data.reachEventCB);
     REQUIRE(data.reachInstCB);
+    REQUIRE(data.reachInstrumentCB);
 }
 
 TEST_CASE("VMTest-MoveAssignmentOperator") {
@@ -755,8 +781,8 @@ TEST_CASE("VMTest-MoveAssignmentOperator") {
     QBDI::VM* vm2 = vm2_.vm.get();
     REQUIRE( vm1 != vm2 );
 
-    MoveCallbackStruct data1 {vm1, true, false, false};
-    MoveCallbackStruct data2 {vm2, true, false, false};
+    MoveCallbackStruct data1 {vm1, true, false, false, false, false};
+    MoveCallbackStruct data2 {vm2, true, false, false, false, false};
 
     bool instrumented = vm1->addInstrumentedModuleFromAddr((QBDI::rword)&dummyFunCall);
     REQUIRE(instrumented);
@@ -764,9 +790,11 @@ TEST_CASE("VMTest-MoveAssignmentOperator") {
     REQUIRE(instrumented);
 
     vm1->addCodeCB(QBDI::InstPosition::POSTINST, verifyVMRef, &data1);
+    vm1->addInstrRule(instrumentCopyCB, QBDI::ANALYSIS_INSTRUCTION, &data1);
     vm1->addVMEventCB(QBDI::SEQUENCE_ENTRY | QBDI::SEQUENCE_EXIT | QBDI::BASIC_BLOCK_NEW, allowedNewBlock, &data1);
 
     vm2->addCodeCB(QBDI::InstPosition::POSTINST, verifyVMRef, &data2);
+    vm2->addInstrRule(instrumentCopyCB, QBDI::ANALYSIS_INSTRUCTION, &data2);
     vm2->addVMEventCB(QBDI::SEQUENCE_ENTRY | QBDI::SEQUENCE_EXIT | QBDI::BASIC_BLOCK_NEW, allowedNewBlock, &data2);
 
     QBDI::rword retvalue;
@@ -775,18 +803,22 @@ TEST_CASE("VMTest-MoveAssignmentOperator") {
     REQUIRE(retvalue == 350);
     REQUIRE(data1.reachEventCB);
     REQUIRE(data1.reachInstCB);
+    REQUIRE(data1.reachInstrumentCB);
 
     data1.reachEventCB = false;
     data1.reachInstCB = false;
+    data1.reachInstrumentCB = false;
     data1.allowedNewBlock = false;
 
     vm2->call(&retvalue, (QBDI::rword) dummyFun1, {670});
     REQUIRE(retvalue == 670);
     REQUIRE(data2.reachEventCB);
     REQUIRE(data2.reachInstCB);
+    REQUIRE(data2.reachInstrumentCB);
 
     data2.reachEventCB = false;
     data2.reachInstCB = false;
+    data2.reachInstrumentCB = false;
     data2.allowedNewBlock = false;
 
     data1.expectedRef = vm2;
@@ -800,11 +832,18 @@ TEST_CASE("VMTest-MoveAssignmentOperator") {
     vm1 = nullptr;
 
     vm2->call(&retvalue, (QBDI::rword) dummyFun1, {780});
+
     REQUIRE(retvalue == 780);
     REQUIRE(data1.reachEventCB);
     REQUIRE(data1.reachInstCB);
-    REQUIRE(( not data2.reachEventCB));
-    REQUIRE(( not data2.reachInstCB));
+    REQUIRE_FALSE(data1.reachInstrumentCB);
+    REQUIRE_FALSE(data2.reachEventCB);
+    REQUIRE_FALSE(data2.reachInstCB);
+    REQUIRE_FALSE(data2.reachInstrumentCB);
+
+    data1.allowedNewBlock = true;
+    vm2->precacheBasicBlock((QBDI::rword) dummyFun0);
+    REQUIRE(data1.reachInstrumentCB);
 }
 
 TEST_CASE("VMTest-CopyAssignmentOperator") {
@@ -815,8 +854,8 @@ TEST_CASE("VMTest-CopyAssignmentOperator") {
     QBDI::VM* vm2 = vm2_.vm.get();
     REQUIRE( vm1 != vm2 );
 
-    MoveCallbackStruct data1 {vm1, true, false, false};
-    MoveCallbackStruct data2 {vm2, true, false, false};
+    MoveCallbackStruct data1 {vm1, true, false, false, false, false};
+    MoveCallbackStruct data2 {vm2, true, false, false, false, false};
 
     bool instrumented = vm1->addInstrumentedModuleFromAddr((QBDI::rword)&dummyFunCall);
     REQUIRE(instrumented);
@@ -825,9 +864,11 @@ TEST_CASE("VMTest-CopyAssignmentOperator") {
 
     vm1->addCodeCB(QBDI::InstPosition::POSTINST, verifyVMRef, &data1);
     vm1->addCodeCB(QBDI::InstPosition::POSTINST, verifyCB2, &data1);
+    vm1->addInstrRule(instrumentCopyCB, QBDI::ANALYSIS_INSTRUCTION, &data1);
     vm1->addVMEventCB(QBDI::SEQUENCE_ENTRY | QBDI::SEQUENCE_EXIT | QBDI::BASIC_BLOCK_NEW, allowedNewBlock, &data1);
 
     vm2->addCodeCB(QBDI::InstPosition::POSTINST, verifyVMRef, &data2);
+    vm2->addInstrRule(instrumentCopyCB, QBDI::ANALYSIS_INSTRUCTION, &data2);
     vm2->addVMEventCB(QBDI::SEQUENCE_ENTRY | QBDI::SEQUENCE_EXIT | QBDI::BASIC_BLOCK_NEW, allowedNewBlock, &data2);
 
     QBDI::rword retvalue;
@@ -836,10 +877,12 @@ TEST_CASE("VMTest-CopyAssignmentOperator") {
     REQUIRE(retvalue == 350);
     REQUIRE(data1.reachEventCB);
     REQUIRE(data1.reachInstCB);
+    REQUIRE(data1.reachInstrumentCB);
     REQUIRE(data1.reachCB2);
 
     data1.reachEventCB = false;
     data1.reachInstCB = false;
+    data1.reachInstrumentCB = false;
     data1.allowedNewBlock = false;
     data1.reachCB2 = false;
 
@@ -847,10 +890,12 @@ TEST_CASE("VMTest-CopyAssignmentOperator") {
     REQUIRE(retvalue == 670);
     REQUIRE(data2.reachEventCB);
     REQUIRE(data2.reachInstCB);
-    REQUIRE(( not data2.reachCB2));
+    REQUIRE(data2.reachInstrumentCB);
+    REQUIRE_FALSE(data2.reachCB2);
 
     data2.reachEventCB = false;
     data2.reachInstCB = false;
+    data2.reachInstrumentCB = false;
     data2.allowedNewBlock = false;
     data2.expectedRef = nullptr;
 
@@ -861,23 +906,29 @@ TEST_CASE("VMTest-CopyAssignmentOperator") {
     REQUIRE(retvalue == 780);
     REQUIRE(data1.reachEventCB);
     REQUIRE(data1.reachInstCB);
+    REQUIRE_FALSE(data1.reachInstrumentCB);
     REQUIRE(data1.reachCB2);
-    REQUIRE(( not data2.reachEventCB));
-    REQUIRE(( not data2.reachInstCB));
-    REQUIRE(( not data2.reachCB2));
+    REQUIRE_FALSE(data2.reachEventCB);
+    REQUIRE_FALSE(data2.reachInstCB);
+    REQUIRE_FALSE(data2.reachCB2);
+    REQUIRE_FALSE(data2.reachInstrumentCB);
 
     data1.reachEventCB = false;
     data1.reachInstCB = false;
+    data1.reachInstrumentCB = false;
     data1.allowedNewBlock = true;
     data1.expectedRef = vm2;
     data1.reachCB2 = false;
 
     vm2->call(&retvalue, (QBDI::rword) dummyFun1, {567});
+
     REQUIRE(retvalue == 567);
     REQUIRE(data1.reachEventCB);
     REQUIRE(data1.reachInstCB);
+    REQUIRE(data1.reachInstrumentCB);
     REQUIRE(data1.reachCB2);
-    REQUIRE(( not data2.reachEventCB));
-    REQUIRE(( not data2.reachInstCB));
-    REQUIRE(( not data2.reachCB2));
+    REQUIRE_FALSE(data2.reachEventCB);
+    REQUIRE_FALSE(data2.reachInstCB);
+    REQUIRE_FALSE(data2.reachCB2);
+    REQUIRE_FALSE(data2.reachInstrumentCB);
 }